+++ test/zdtm/static/unlink_fstat02.c
@@ -47,7 +47,8 @@ int main(int argc, char ** argv)
 	}
 
 	if (fst.st_size != 0) {
-		pr_perror("%s file size eq %ld", filename, fst.st_size);
+		pr_perror("%s file size eq %lld",
+				filename, (long long)fst.st_size);
 		goto failed;
 	}
 
@@ -86,13 +87,15 @@ int main(int argc, char ** argv)
 	}
 
 	if (fst.st_size != fsize) {
-		fail("(via fstat): file size changed to %ld", fst.st_size);
+		fail("(via fstat): file size changed to %lld",
+				(long long)fst.st_size);
 		goto failed;
 	}
 
 	fst.st_size = lseek(fd[0], 0, SEEK_END);
 	if (fst.st_size != fsize) {
-		fail("(via lseek): file size changed to %ld", fst.st_size);
+		fail("(via lseek): file size changed to %lld",
+				(long long)fst.st_size);
 		goto failed;
 	}
